refactor(EditRole): decomposing
